expand OWs to u16 (step 1/N)
first commit.
followed Deokishisu's commit "Expand OWs to u16" in the CrystalDust repo.

compiles and works almost correctly.
All overworlds are set to "Red normal" (0 index), no idea why.

Did not test going beyond 255 yet.

+++ src/event_object_movement.c
@@ -1380,7 +1380,7 @@ static bool8 sub_805E238(struct ObjectEventTemplate *template, u8 var, s16 x, s1
 
 static bool8 sub_805E27C(struct ObjectEventTemplate *template, s16 x, s16 y)
 {
-    if ((u8)(template->graphicsId - OBJ_EVENT_GFX_CUT_TREE) > 1)
+    if ((u16)(template->graphicsId - OBJ_EVENT_GFX_CUT_TREE) > 1)
         return TRUE;
 
     if (gSaveBlock1Ptr->pos.x < x)
@@ -1413,7 +1413,7 @@ static bool8 sub_805E2E8(struct ObjectEventTemplate *template, s16 x, s16 y)
     x2 = VMap.Xsize - 16;
     y2 = VMap.Ysize - 15;
     
-    if ((u8)(template->graphicsId - OBJ_EVENT_GFX_CUT_TREE) > 1)
+    if ((u16)(template->graphicsId - OBJ_EVENT_GFX_CUT_TREE) > 1)
         return TRUE;
     
     if (!gSaveBlock1Ptr->pos.x)
@@ -1630,7 +1630,7 @@ u8 SpawnSpecialObjectEvent(struct ObjectEventTemplate *objectEventTemplate)
     return TrySpawnObjectEventTemplate(objectEventTemplate, gSaveBlock1Ptr->location.mapNum, gSaveBlock1Ptr->location.mapGroup, cameraX, cameraY);
 }
 
-int SpawnSpecialObjectEventParameterized(u8 graphicsId, u8 movementBehavior, u8 localId, s16 x, s16 y, u8 z)
+int SpawnSpecialObjectEventParameterized(u16 graphicsId, u8 movementBehavior, u8 localId, s16 x, s16 y, u8 z)
 {
     struct ObjectEventTemplate objectEventTemplate;
 
@@ -1715,7 +1715,7 @@ u8 AddPseudoObjectEvent(u16 graphicsId, SpriteCallback callback, s16 x, s16 y, u
     return spriteId;
 }
 
-u8 sprite_new(u8 graphicsId, u8 a1, s16 x, s16 y, u8 z, u8 direction)
+u8 sprite_new(u16 graphicsId, u8 a1, s16 x, s16 y, u8 z, u8 direction)
 {
     u8 spriteId;
     struct Sprite *sprite;
@@ -1757,7 +1757,7 @@ u8 sprite_new(u8 graphicsId, u8 a1, s16 x, s16 y, u8 z, u8 direction)
     return spriteId;
 }
 
-u8 sub_805EB44(u8 graphicsId, u8 a1, s16 x, s16 y)
+u8 sub_805EB44(u16 graphicsId, u8 a1, s16 x, s16 y)
 {
     u8 spriteId;
     struct Sprite *sprite;
@@ -1962,7 +1962,7 @@ static void SetPlayerAvatarObjectEventIdAndObjectId(u8 objectEventId, u8 spriteI
     SetPlayerAvatarExtraStateTransition(gObjectEvents[objectEventId].graphicsId, PLAYER_AVATAR_FLAG_CONTROLLABLE);
 }
 
-void ObjectEventSetGraphicsId(struct ObjectEvent *objectEvent, u8 graphicsId)
+void ObjectEventSetGraphicsId(struct ObjectEvent *objectEvent, u16 graphicsId)
 {
     const struct ObjectEventGraphicsInfo *graphicsInfo;
     struct Sprite *sprite;
@@ -2015,7 +2015,7 @@ void ObjectEventSetGraphicsId(struct ObjectEvent *objectEvent, u8 graphicsId)
     }
 }
 
-void ObjectEventSetGraphicsIdByLocalIdAndMap(u8 localId, u8 mapNum, u8 mapGroup, u8 graphicsId)
+void ObjectEventSetGraphicsIdByLocalIdAndMap(u8 localId, u8 mapNum, u8 mapGroup, u16 graphicsId)
 {
     u8 objectEventId;
 
@@ -2050,7 +2050,7 @@ void PlayerObjectTurn(struct PlayerAvatar *playerAvatar, u8 direction)
     ObjectEventTurn(&gObjectEvents[playerAvatar->objectEventId], direction);
 }
 
-const struct ObjectEventGraphicsInfo *GetObjectEventGraphicsInfo(u8 graphicsId)
+const struct ObjectEventGraphicsInfo *GetObjectEventGraphicsInfo(u16 graphicsId)
 {
     if (graphicsId >= OBJ_EVENT_GFX_VARS)
         graphicsId = VarGetObjectEventGraphicsId(graphicsId - OBJ_EVENT_GFX_VARS);
@@ -9263,7 +9263,7 @@ void TurnObjectEvent(u8 objectEventId, u8 direction)
     }
 }
 
-void RfuUnionObjectSetFacingDirection(u8 objectEventId, u8 direction)
+void RfuUnionObjectSetFacingDirection(u8 objectEventId, u16 direction)
 {
     u8 animNum;
     int spriteId = GetObjectEventSpriteId(objectEventId);
